rename createcomponent to recipecrud
